+++ errors.h
@@ -8,7 +8,7 @@
 
 typedef struct {
 	int type;
-	long number;
+	double number;
 	int error;
 } lisp_value;
 
@@ -18,7 +18,7 @@ enum { LISP_VALUE_NUMBER, LISP_VALUE_ERROR };
 //Create enumeration of possible error types
 enum{ LISP_VALUE_DIVIDE_ZERO, LISP_VALUE_BAD_OPERATOR, LISP_VALUE_BAD_NUMBER};
 
-lisp_value lisp_value_number(long x);
+lisp_value lisp_value_number(double x);
 lisp_value lisp_value_error(int x);
 void print_lisp_value(lisp_value v);
 void print_lisp_value_newline(lisp_value v);
